+++ Player.cpp
@@ -1,7 +1,7 @@
 #include "Player.hpp"
 
 
-Player::Player(int x, int y, int width, int height, int img_width, int img_height, int speed, Direction direction, char* image_path, SDL_Renderer* renderer) 
+Player::Player(int x, int y, int width, int height, int img_width, int img_height, double speed, Direction direction, char* image_path, SDL_Renderer* renderer) 
 : GameObject(x, y, width, height, img_width, img_height, speed, direction, image_path, renderer) {
     this->timer_retain = 0;
     this->frame_number = 0;
