+++ Projectile.cpp
@@ -1,7 +1,7 @@
 #include "Projectile.hpp"
 
 
-Projectile::Projectile(int x, int y, int width, int height, int img_width, int img_height, int speed, Direction direction, char* image_path, SDL_Renderer* renderer) 
+Projectile::Projectile(int x, int y, int width, int height, int img_width, int img_height, double speed, Direction direction, char* image_path, SDL_Renderer* renderer) 
 : GameObject(x, y, width, height, img_width, img_height, speed, direction, image_path, renderer) {
     this->timer_retain = 0;
     this->frame_number = 0;
